@@ -25,49 +25,52 @@
   },
   "homepage": "https://github.com/OpenSpace/OpenSpace#readme",
   "dependencies": {
-    "core-js": "^3.1.3",
-    "lodash": "^4.17.15",
-    "openspace-api-js": "^0.1.2",
+    "core-js": "^3.11.2",
+    "d3-scale": "^3.3.0",
+    "lodash": "^4.17.21",
+    "openspace-api-js": "^0.1.5",
     "prop-types": "^15.7.2",
-    "re-resizable": "^5.0.1",
-    "react": "^16.8.6",
-    "react-color": "^2.17.3",
-    "react-dom": "^16.8.6",
-    "react-draggable": "^3.3.0",
-    "react-hot-loader": "^4.11.0",
-    "react-input-autosize": "^2.2.1",
-    "react-redux": "^7.1.0",
-    "react-router-dom": "^5.0.1",
-    "react-select": "^3.0.4",
-    "redux": "^4.0.1",
-    "shallow-equal": "^1.2.0",
+    "re-resizable": "^6.9.0",
+    "react": "^16.14.0",
+    "react-color": "^2.19.3",
+    "react-dom": "^16.14.0",
+    "react-draggable": "^4.4.3",
+    "react-hot-loader": "^4.13.0",
+    "react-input-autosize": "^2.2.2",
+    "react-redux": "^7.2.4",
+    "react-router-dom": "^5.2.0",
+    "react-select": "^3.2.0",
+    "react-simple-keyboard": "^3.2.72",
+    "redux": "^4.1.0",
+    "shallow-equal": "^1.2.1",
+    "simple-keyboard-layouts": "^3.0.176",
     "svg-react-loader": "^0.4.6"
   },
   "devDependencies": {
-    "@babel/core": "^7.4.5",
-    "@babel/preset-env": "^7.4.5",
-    "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^8.0.6",
-    "copyfiles": "^2.1.0",
-    "cross-env": "^5.2.0",
-    "css-loader": "^3.0.0",
+    "@babel/core": "^7.14.0",
+    "@babel/preset-env": "^7.14.1",
+    "@babel/preset-flow": "^7.13.13",
+    "@babel/preset-react": "^7.13.13",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.2.2",
+    "copyfiles": "^2.4.1",
+    "cross-env": "^5.2.1",
+    "css-loader": "^3.6.0",
     "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-import": "^2.17.3",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.13.0",
-    "file-loader": "^4.0.0",
-    "flow-bin": "^0.101.0",
-    "node-sass": "^4.12.0",
-    "rimraf": "^2.6.3",
-    "sass-loader": "^7.1.0",
+    "eslint-config-airbnb": "^17.1.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.23.2",
+    "file-loader": "^4.3.0",
+    "flow-bin": "^0.101.1",
+    "node-sass": "^4.14.1",
+    "rimraf": "^2.7.1",
+    "sass-loader": "^7.3.1",
     "style-loader": "^0.23.1",
-    "url-loader": "^2.0.0",
-    "webpack": "^4.34.0",
-    "webpack-cli": "^3.3.4",
-    "webpack-dev-server": "^3.7.1"
+    "url-loader": "^2.3.0",
+    "webpack": "^4.46.0",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.2"
   },
   "babel": {
     "presets": [
