@@ -4,8 +4,10 @@ import { excludeKeys } from '../../../../utils/helpers';
 import styles from './NumericInput.scss';
 import Input from '../Input/Input';
 import Tooltip from '../../Tooltip/Tooltip';
-import { round10 } from '../../../../utils/rounding';
+import { roundValueToStepSize } from '../../../../utils/rounding';
 import { clamp } from 'lodash/number';
+
+const Scale = require('d3-scale');
 
 class NumericInput extends Component {
   constructor(props) {
@@ -15,27 +17,99 @@ class NumericInput extends Component {
       value: props.value,
       showTextInput: false,
       id: `numericinput-${Input.nextId}`,
-      hoverHint: null,
+      hoverHint: null
     };
 
-    this.onChange = this.onChange.bind(this);
-    this.onTextBlur = this.onTextBlur.bind(this);
+    this.roundValueToStepSize = this.roundValueToStepSize.bind(this);
+
+    this.updateSliderScale = this.updateSliderScale.bind(this);
+    this.valueToSliderPos = this.valueToSliderPos.bind(this);
+    this.valueFromSliderPos = this.valueFromSliderPos.bind(this);
+
     this.onHover = this.onHover.bind(this);
     this.onLeave = this.onLeave.bind(this);
+    this.onTextBlur = this.onTextBlur.bind(this);
+    this.onSliderChange = this.onSliderChange.bind(this);
     this.enableTextInput = this.enableTextInput.bind(this);
     this.disableTextInput = this.disableTextInput.bind(this);
+    this.updateValue = this.updateValue.bind(this);
+
+    this.sliderResolution = 10000;
+    this.scale = this.updateSliderScale();
   }
 
-  componentWillReceiveProps({ value }) {
-    this.setState({ value });
+  componentDidUpdate(prevProps, prevState) {
+    // Update state value variable when we get new props
+    if (prevProps.value !== this.props.value) {
+      this.setState({ value: this.props.value });
+    }
+
+    const scaleNeedsUpdate = (prevProps.min !== this.props.min) ||
+                             (prevProps.max !== this.props.max) ||
+                             (prevProps.step !== this.props.step) ||
+                             (prevProps.exponent !== this.props.exponent);
+
+    if (scaleNeedsUpdate) {
+      this.scale = this.updateSliderScale();
+    }
+  }
+
+  roundValueToStepSize(value) {
+    return roundValueToStepSize(value, this.props.step);
+  }
+
+  updateSliderScale() {
+    const {exponent, min, max, step} = this.props;
+
+    // Prevent setting exponent to zero, as it breaks the scale
+    const exp = (exponent == 0) ? 1.0 : exponent;
+
+    // If linear scale, we want the resolution to match the step size
+    if (exp == 1.0) {
+      const nSteps = Math.ceil((max - min) / step);
+      this.sliderResolution = nSteps;
+    }
+
+    // The slider is logarithmic, but the scaling of the value increases exponentially
+    return Scale.scalePow()
+            .exponent(exp)
+            .domain([0, this.sliderResolution]) // slider pos
+            .range([min, max]); // allowed values
+  }
+
+  valueToSliderPos(value) {
+    return this.scale.invert(value);
+  }
+
+  valueFromSliderPos(sliderValue) {
+    const scaledValue = this.scale(sliderValue);
+    // If almost max, return max, as rounding will prevent this if the step size
+    // is not well chosen
+    if (scaledValue > 0.999 * this.props.max) {
+      return this.props.max;
+    }
+    return this.roundValueToStepSize(scaledValue);
   }
 
   /**
    * callback for input
    * @param event InputEvent
    */
-  onChange(event) {
+  onSliderChange(event) {
+    const sliderValue = Number.parseFloat(event.currentTarget.value);
+    const newValue = this.valueFromSliderPos(sliderValue);
+    this.updateValue(newValue);
+  }
+
+  onTextBlur(event) {
     const value = Number.parseFloat(event.currentTarget.value);
+    if(!isNaN(value)) {
+      this.updateValue(value);
+    }
+    this.disableTextInput();
+  }
+
+  updateValue(value) {
     const { max, min } = this.props;
 
     if (value > max || value < min) return;
@@ -44,12 +118,7 @@ class NumericInput extends Component {
     this.setState({ value });
 
     // send to the onChange (if any)!
-    this.props.onChange(event);
-  }
-
-  onTextBlur(event) {
-    this.onChange(event);
-    this.disableTextInput();
+    this.props.onValueChanged(value);
   }
 
   onHover(event) {
@@ -73,7 +142,7 @@ class NumericInput extends Component {
   }
 
   enableTextInput(event) {
-    if (this.props.disableInput) {
+    if (this.props.disabled) {
       return;
     }
     this.setState({ showTextInput: true, hoverHint: null });
@@ -89,7 +158,7 @@ class NumericInput extends Component {
     if (this.showTextInput) {
       return (
         <Input
-          {...excludeKeys(this.props, 'reverse onChange disableInput inputOnly noHoverHint noTooltip')}
+          {...excludeKeys(this.props, 'reverse onValueChanged inputOnly noHoverHint noTooltip noValue exponent')}
           type="number"
           value={value}
           onBlur={this.onTextBlur}
@@ -99,11 +168,17 @@ class NumericInput extends Component {
       );
     }
 
-    const { placeholder, className, label, wide, reverse, min, max, noValue } = this.props;
-    const doNotInclude = 'wide reverse onChange value className type ' +
-                         'disableInput inputOnly label noHoverHint noTooltip noValue';
+    const { placeholder, className, label, wide, reverse, noValue } = this.props;
+    const doNotInclude = 'wide reverse onValueChanged value className type min max step exponent ' +
+                         'inputOnly label noHoverHint noTooltip noValue';
     const inheritedProps = excludeKeys(this.props, doNotInclude);
     const hoverHintOffset = reverse ? 1 - hoverHint : hoverHint;
+
+    const sliderValue = this.valueToSliderPos(value);
+
+    // HoverHintOffset is in [0, 1]. Scale to full slider range
+    const scaledHoverHintOffset = this.sliderResolution * hoverHintOffset;
+    const tooltipValue = this.valueFromSliderPos(scaledHoverHintOffset);
 
     return (
       <div
@@ -116,17 +191,20 @@ class NumericInput extends Component {
         )}
         { !this.props.noTooltip && hoverHint !== null && (
           <Tooltip style={{ left: `${100 * hoverHint}%` }}>
-            { round10(min + hoverHintOffset * (max - min), Math.log10(this.props.step)) }
+            { tooltipValue }
           </Tooltip>
         )}
         <input
           {...inheritedProps}
           id={id}
           type="range"
-          value={value}
+          value={sliderValue}
+          min={0}
+          max={this.sliderResolution}
+          step={1}
           className={`${className} ${styles.range}`}
-          style={{ '--min': min, '--max': max, '--value': value, direction: reverse ? "rtl" : "ltr" }}
-          onChange={this.onChange}
+          style={{ '--min': 0, '--max': this.sliderResolution, '--value': sliderValue, direction: reverse ? "rtl" : "ltr" }}
+          onChange={this.onSliderChange}
           onMouseMove={this.onHover}
           onMouseLeave={this.onLeave}
         />
@@ -144,7 +222,7 @@ class NumericInput extends Component {
 NumericInput.propTypes = {
   className: PropTypes.string,
   disabled: PropTypes.bool,
-  disableInput: PropTypes.bool,
+  exponent: PropTypes.number,
   inputOnly: PropTypes.bool,
   label: PropTypes.node,
   max: PropTypes.number,
@@ -153,7 +231,7 @@ NumericInput.propTypes = {
   noHoverHint: PropTypes.bool,
   noTooltip: PropTypes.bool,
   noValue: PropTypes.bool,
-  onChange: PropTypes.func,
+  onValueChanged: PropTypes.func,
   placeholder: PropTypes.string.isRequired,
   step: PropTypes.number,
   value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
@@ -163,7 +241,7 @@ NumericInput.propTypes = {
 NumericInput.defaultProps = {
   className: '',
   disabled: false,
-  disableInput: false,
+  exponent: 1.0,
   inputOnly: false,
   label: null,
   max: 100,
@@ -172,7 +250,7 @@ NumericInput.defaultProps = {
   noHoverHint: false,
   noTooltip: false,
   noValue: false,
-  onChange: () => {},
+  onValueChanged: () => {},
   step: 1,
   value: 0,
   wide: true,
