@@ -209,7 +209,6 @@ export const selectImgSkyBrowser = data => ({
   payload: data
 });
 
-
 export const subscribeToShortcuts = () => ({
   type: actionTypes.subscribeToShortcuts,
   payload: {},
@@ -225,10 +224,21 @@ export const initializeShortcuts = data => ({
   payload: data
 });
 
-export const executeShortcut = index => ({
-  type: actionTypes.executeShortcut,
-  payload: index
+export const triggerAction = data => ({
+  type: actionTypes.triggerAction,
+  payload: data
 });
+
+export const setActionsPath = data => ({
+  type: actionTypes.setActionsPath,
+  payload: data
+});
+
+export const toggleKeybindViewer = data => ({
+  type: actionTypes.toggleKeybindViewer,
+  payload: data
+});
+
 
 export const setNavigationAction = data => ({
   type: actionTypes.setNavigationAction,
