@@ -3,6 +3,7 @@ import NumericInput from '../../common/Input/NumericInput/NumericInput';
 import InfoBox from '../../common/InfoBox/InfoBox';
 import { connectProperty } from './connectProperty';
 import { copyTextToClipboard } from '../../../utils/helpers';
+import styles from './Property.scss';
 
 class NumericProperty extends Component {
   constructor(props) {
@@ -28,9 +29,8 @@ class NumericProperty extends Component {
     return this.props.description.MetaData.isReadOnly;
   }
 
-  onChange(event) {
-    const { value } = event.currentTarget;
-    this.props.dispatcher.set(parseFloat(value));
+  onChange(newValue) {
+    this.props.dispatcher.set(newValue);
   }
 
   get descriptionPopup() {
@@ -46,18 +46,21 @@ class NumericProperty extends Component {
 
   render() {
     const { description, value } = this.props;
-    const { SteppingValue, MaximumValue, MinimumValue } = description.AdditionalData;
+    const { SteppingValue, MaximumValue, MinimumValue, Exponent } = description.AdditionalData;
     return (
-      <NumericInput
-        value={value}
-        label={(<span onClick={this.copyUri}>{description.Name} {this.descriptionPopup}</span>)}
-        placeholder={description.Name}
-        onChange={this.onChange}
-        step={SteppingValue}
-        max={MaximumValue}
-        min={MinimumValue}
-        disabled={this.disabled}
-      />
+      <div className={`${this.disabled ? styles.disabled : ''}`}>
+        <NumericInput
+          value={value}
+          label={(<span onClick={this.copyUri}>{description.Name} {this.descriptionPopup}</span>)}
+          placeholder={description.Name}
+          onValueChanged={this.onChange}
+          step={SteppingValue}
+          exponent={Exponent}
+          max={MaximumValue}
+          min={MinimumValue}
+          disabled={this.disabled}
+        />
+      </div>
     );
   }
 }
