@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import { excludeKeys } from '../../../../utils/helpers';
 import styles from './Checkbox.scss';
 import MaterialIcon from '../../MaterialIcon/MaterialIcon';
-import Input from '../Input/Input';
 
 class Checkbox extends Component {
   constructor(props) {
@@ -11,17 +10,19 @@ class Checkbox extends Component {
     this.onClick = this.onClick.bind(this);
   }
 
-
   onClick(evt) {
+    if(this.props.disabled) {
+      return;
+    }
     this.props.setChecked(!this.props.checked);
     evt.stopPropagation();
   }
 
   render() {
-    const { checked, label, wide, left } = this.props;
+    const { checked, label, wide, left, className } = this.props;
 
     return (
-      <div onClick={this.onClick} className={`${styles.wrapper} ${wide ? styles.wide : ''} ${left ? styles.left : ''}`}>
+      <div onClick={this.onClick} className={`${styles.wrapper} ${className} ${wide ? styles.wide : ''} ${left ? styles.left : ''}`}>
         <MaterialIcon className={styles.checkbox} icon={checked ? "check_box" : "check_box_outline_blank"} />
         { label && <label>{ label }</label> }
       </div>
@@ -31,6 +32,7 @@ class Checkbox extends Component {
 
 Checkbox.propTypes = {
   checked: PropTypes.bool,
+  disabled: PropTypes.bool,
   setChecked: PropTypes.func,
   label: PropTypes.node,
   left: PropTypes.bool,
