@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { createStore } from 'redux';
+import { createStore, compose } from 'redux';
 // AppContainer is a necessary wrapper component for hot module reloading
 import { AppContainer } from 'react-hot-loader';
 import openspaceApp from './api/Reducers';
@@ -10,9 +10,13 @@ import App from './App';
 
 /* global document */
 
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;    // Redux DevTools Extension courtesy of https://github.com/zalmoxisus/redux-devtools-extension
+
 const store = createStore(
   openspaceApp,
-  middleware,
+  composeEnhancers(
+    middleware,
+  )
 );
 
 const render = (Component) => {
